target3 modify shuffle method

+++ utils/common_utils.py
@@ -49,7 +49,7 @@ def shuffle_image(img, blockLen):
     width, height = img.size
     xblock = width / blockLen
     yblock = height / blockLen
-    blockmap = [((int)xb*blockLen, (int)yb*blockLen, (xb+1)*blockLen, (yb+1)*blockLen)
+    blockmap = [((int)(xb)*blockLen, (int)(yb)*blockLen, ((int)(xb)+1)*blockLen, ((int)(yb)+1)*blockLen)
         for xb in range(xblock) for yb in range(yblock)]
 
     shuffle = list(blockmap)
